requirements.txt
Django==1.7.7
Pillow==2.8.0
app-version==0.2.1
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
django-allauth==0.19.1
django-appconf==1.0.1
django-audit-log==0.7.0
django-bootstrap3==5.2.0
django-braces==1.4.0
django-debug-toolbar==1.3.0
django-extra-views==0.6.5
django-permission==0.8.5
django-suit==0.2.13
django-tagging==0.3.4
django-toolbelt==0.0.1
djrill==1.3.0
gunicorn==19.3.0
lxml==3.4.2
oauthlib==0.7.2
psycopg2==2.6
python3-openid==3.0.5
requests==2.6.0
requests-oauthlib==0.4.2
six==1.9.0
sorl-thumbnail==12.2
sqlparse==0.1.14
static3==0.5.1
swapper==0.2.1
tolerance==0.1.2
whitenoise==1.0.6
